chore: safe dashboard added
