+++ graphql/getWorkspaceMembers.graphql
@@ -15,6 +15,11 @@ query getWorkspaceMembers($actorId: Bytes!) {
       logoIpfsHash
       title
       supportedNetworks
+      safe {
+        id
+        chainId
+        address
+      }
       tokens {
         address
         label
