+++ src/v2/components/NetworkTransactionModal.tsx
@@ -1,5 +1,6 @@
 import { ReactNode } from 'react'
 import { AlertDialogOverlay, Box, Button, Divider, Flex, HStack, Image, Modal, ModalBody, ModalContent, ModalHeader, Text, VStack } from '@chakra-ui/react'
+import { useRouter } from 'next/router'
 import { CheckCircle } from 'src/v2/assets/custom chakra icons/CheckCircle'
 import { ExternalLink } from 'src/v2/assets/custom chakra icons/ExternalLink'
 
@@ -35,7 +36,7 @@ export default ({
 	viewLink,
 	onClose,
 }: NetworkTransactionModalProps) => {
-
+	const router = useRouter()
 	return (
 		<Modal
 			isOpen={isOpen}
@@ -131,7 +132,12 @@ export default ({
 										isDisabled={currentStepIndex < steps.length}
 										ml='auto'
 										variant='primaryV2'
-										onClick={onClose}>
+										onClick={
+											() => {
+												onClose()
+												router.reload()
+											}
+										}>
 										Okay
 									</Button>
 								</Flex>
